feat: centralize render frame modes
