+++ readme.md
@@ -10,20 +10,19 @@ Minimal Node + browser setup that:
 
 ## Architecture
 - `bin/engine.mjs` launches the HTTP server and invokes the engine's `start` function, streaming NDJSON frames.
-- `src/render-scene.mjs` implements shared scene rendering and post-processing.
-- `src/engine.mjs` renders frames and exposes live parameters.
+- `src/render-scene.mjs` implements shared scene rendering, post-processing and the `renderFrames` helper used by both the engine and UI.
+- `src/engine.mjs` runs the render loop with `renderFrames` and exposes live parameters.
 - `src/server.mjs` exports a `startServer` helper that serves the UI and relays WebSocket param updates.
 - `src/ui/` contains the browser preview and controls.
 
 Runtime parameters are grouped under `effects` for effect-specific settings
 and `post` for modifiers like brightness, tint and strobe which can be applied on top.
-A `renderMode` parameter controls how scenes map to the two walls: `duplicate` renders once for both, `extended` renders a double-width scene and splits it, and `mirror` flips the scene for the right wall.
+A `renderMode` parameter controls how scenes map to the two walls: `duplicate` renders once for both, `extended` renders a double-width scene and splits it, and `mirror` flips the scene for the right wall. Both the engine and browser preview rely on `renderFrames` to handle these modes.
 
 ## Frame pipeline
-1. The engine renders the active effect into a floating point RGB buffer (`leftFrame`).
-2. Post-processing modifiers run on that buffer and the result is combined into `rightFrame` according to `renderMode`.
-3. Each frame is sliced according to the configured layouts and emitted as base64-encoded `rgb8` NDJSON.
-4. The browser preview reuses these frame buffers to draw the scene and per-LED indicators.
+1. `renderFrames` draws the active effect, applies post-processing and fills the `leftFrame` and `rightFrame` buffers based on `renderMode`.
+2. Each frame is sliced according to the configured layouts and emitted as base64-encoded `rgb8` NDJSON.
+3. The browser preview reuses these frame buffers to draw the scene and per-LED indicators.
 
 ## Quick start
 1. Open your terminal
