+++ src/readme.md
@@ -2,8 +2,8 @@
 
 Core runtime code for BarnLights Playbox:
 
-- `render-scene.mjs` – shared scene rendering and post-processing helpers.
-- `engine.mjs` – side‑effect‑free render loop exposing `params` (including `renderMode`); call `start()` to emit SLICES_NDJSON.
+- `render-scene.mjs` – shared scene rendering, post-processing and `renderFrames` helper for mapping scenes to both walls.
+- `engine.mjs` – side‑effect‑free render loop using `renderFrames` and exposing `params` (including `renderMode`); call `start()` to emit SLICES_NDJSON.
 - `server.mjs` – HTTP/WebSocket server serving the UI and applying param updates.
 - `config-store.mjs` – read/write helpers for saving and loading effect presets.
 - `effects/` – effect implementations, registry and post-processing helpers.
