+++ src/render-scene.mjs
@@ -14,3 +14,38 @@ export function renderScene(sceneF32, sceneW, sceneH, t, P){
     fn(sceneF32, t, post, sceneW, sceneH);
   }
 }
+
+let extendedFrame = null;
+
+// renderFrames: fill left/right frame buffers according to renderMode
+export function renderFrames(leftFrame, rightFrame, P, t){
+  const mode = P.renderMode;
+  if (mode === "extended"){
+    const W = SCENE_W * 2;
+    if (!extendedFrame || extendedFrame.length !== W * SCENE_H * 3){
+      extendedFrame = new Float32Array(W * SCENE_H * 3);
+    }
+    renderScene(extendedFrame, W, SCENE_H, t, P);
+    for (let y = 0; y < SCENE_H; y++){
+      const src = y * W * 3;
+      const dst = y * SCENE_W * 3;
+      leftFrame.set(extendedFrame.subarray(src, src + SCENE_W * 3), dst);
+      rightFrame.set(extendedFrame.subarray(src + SCENE_W * 3, src + W * 3), dst);
+    }
+  } else {
+    renderScene(leftFrame, SCENE_W, SCENE_H, t, P);
+    if (mode === "mirror"){
+      for (let y = 0; y < SCENE_H; y++){
+        for (let x = 0; x < SCENE_W; x++){
+          const src = ((SCENE_H - 1 - y) * SCENE_W + (SCENE_W - 1 - x)) * 3;
+          const dst = (y * SCENE_W + x) * 3;
+          rightFrame[dst] = leftFrame[src];
+          rightFrame[dst + 1] = leftFrame[src + 1];
+          rightFrame[dst + 2] = leftFrame[src + 2];
+        }
+      }
+    } else {
+      rightFrame.set(leftFrame);
+    }
+  }
+}
