+++ src/ui/readme.md
@@ -6,6 +6,6 @@ Browser interface providing live preview and controls.
 - `main.mjs` – entry point for JS logic, wiring modules together, exposes a 'run' function.
 - `connection.mjs` – WebSocket setup and message handling.
 - `controls-logic.mjs` – wires DOM controls to params and renders effect-specific widgets.
-- `preview-renderer.mjs` – scene generation and drawing, duplicating a single scene to both canvases.
+- `renderer.mjs` – uses `renderFrames` to draw the scene for both walls and overlay per-LED indicators.
 - `presets.mjs` – handles saving/retreiving configuration and listing the saved options.
 - `subviews/` – reusable widgets and `renderControls` helper.
